Unit testing

+++ backend/Fakebook/Fakebook.Domain/Repository/UserRepo.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Fakebook.DataAccess.Model;
+using Fakebook.Domain.Extension;
 
 namespace Fakebook.Domain.Repository
 {
@@ -13,16 +14,14 @@ namespace Fakebook.Domain.Repository
     {
         // create a read only for our database
         private readonly FakebookContext _context;
-        public UserRepo(FakebookContext context)
-        {
+        public UserRepo(FakebookContext context) {
             _context = context;
         }
         /// <summary>
         /// Get all users
         /// </summary>
         /// <returns></returns>
-        public IEnumerable<User> GetAllUsers()
-        {
+        public IEnumerable<User> GetAllUsers() {
             var entity = _context.UserEntities
                 .Include(u => u.Followees)
                      .ThenInclude(u => u.Followee)
@@ -37,8 +36,7 @@ namespace Fakebook.Domain.Repository
         /// Get all users asyncronously
         /// </summary>
         /// <returns></returns>
-        public async Task<IEnumerable<User>> GetAllUsersAsync()
-        {
+        public async Task<IEnumerable<User>> GetAllUsersAsync() {
             var entities = await _context.UserEntities
                 .Include(u => u.Followees)
                      .ThenInclude(u => u.Followee)
@@ -55,9 +53,17 @@ namespace Fakebook.Domain.Repository
         /// </summary>
         /// <param name="ids"></param>
         /// <returns></returns>
-        public async Task<IEnumerable<User>> GetUsersByIdsAsync(IEnumerable<int> ids)
-        {
-            var users = await _context.UserEntities
+        public async Task<IEnumerable<User>> GetUsersByIdsAsync(IEnumerable<int> ids) {
+            var userEntities = _context.UserEntities;
+            var userIds = userEntities
+                .Select(u => u.Id)
+                .ToList();
+
+            if (!ids.All(id => userIds.Contains(id))) {
+                throw new ArgumentException("Not all ids requested are present.");
+            }
+
+            var users = await userEntities
                 .Include(u => u.Followees)
                      .ThenInclude(u => u.Followee)
                 .Include(u => u.Followers)
@@ -65,8 +71,7 @@ namespace Fakebook.Domain.Repository
                 .Include(u => u.Posts)
                 .ToListAsync();
 
-            if (!ids.Any() || !users.Any())
-            {
+            if (!ids.Any() || !users.Any()) {
                 return new List<User>();
             }
 
@@ -81,9 +86,14 @@ namespace Fakebook.Domain.Repository
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public async Task<User> GetUserByIdAsync(int id)
-        {
-            var entity = await _context.UserEntities
+        public async Task<User> GetUserByIdAsync(int id) {
+            var entities = _context.UserEntities;
+
+            if (id < 1 || !entities.Any() || id > entities.Max(u => u.Id)) {
+                throw new ArgumentException($"{id} is not a valid id.");
+            }
+
+            var entity = await entities
                 .Where(u => u.Id == id)
                 .Include(u => u.Followees)
                      .ThenInclude(u => u.Followee)
@@ -94,9 +104,16 @@ namespace Fakebook.Domain.Repository
             return user;
         }
 
-        public async Task<User> GetUserByEmailAsync(string email)
-        {
-            var entity = await _context.UserEntities
+        public async Task<User> GetUserByEmailAsync(string email) {
+            var entities = _context.UserEntities;
+
+            email.EnforceEmailCharacters(nameof(email));
+
+            if (!entities.Any()) {
+                throw new ArgumentException($"{email} does not belong to any user");
+            }
+
+            var entity = await entities
                 .Where(e => e.Email == email)
                 .Include(u => u.Followees)
                      .ThenInclude(u => u.Followee)
@@ -127,18 +144,14 @@ namespace Fakebook.Domain.Repository
         /// </summary>
         /// <param name="user"></param>
         /// <returns></returns>
-        public async Task<bool> CreateUser(User user)
-        {
-            var newUser = DbEntityConverter.ToUserEntity(user); // convert
-            try
-            {
+        public async Task<int> CreateUser(User user) {
+            try {
+                var newUser = DbEntityConverter.ToUserEntity(user); // convert
                 await _context.AddAsync(newUser);
                 await _context.SaveChangesAsync();
-                return true;
-            }
-            catch
-            {
-                return false;
+                return newUser.Id;
+            } catch {
+                return -1;
             }
 
         }
@@ -147,50 +160,45 @@ namespace Fakebook.Domain.Repository
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public async Task<bool> DeleteUserAsync(int id)
-        {
-            try
-            {
+        public async Task<bool> DeleteUserAsync(int id) {
+            var entities = _context.UserEntities;
 
-                var entity = await _context.UserEntities.FindAsync(id);
-                _context.UserEntities.Remove(entity);
-                await _context.SaveChangesAsync();
-                return true;
-            }
-            catch
-            {
-                Console.WriteLine("Invalid user");
-                return false;
+            if (id < 1 || !entities.Any() || id > entities.Max(u => u.Id)) {
+                throw new ArgumentException($"{id} is not a valid id.");
             }
 
+            var entity = await entities.FindAsync(id);
+            _context.UserEntities.Remove(entity);
+            await _context.SaveChangesAsync();
+            return true;
         }
         /// <summary>
         /// convert to user entity then save changes.
         /// </summary>
         /// <param name="user"></param>
         /// <returns></returns>
-        public async Task<bool> UpdateUserAsync(int id, User user)
-        {
-            try
-            {
-                UserEntity entity = await _context.UserEntities.FindAsync(id);
-                // assign all the values
-                {
-                    entity.Status = user.Status;
-                    entity.FirstName = user.FirstName;
-                    if (!String.IsNullOrEmpty(user.Email))
-                        entity.Email = user.Email;
-                    entity.LastName = user.LastName;
-                }
-                // save changes.
-                _context.SaveChanges();
-                return true;
-            }
-            catch
-            {
-                return false;
+        public async Task<bool> UpdateUserAsync(int id, User user) {
+            // make sure the user can be converted
+            var userEntity = DbEntityConverter.ToUserEntity(user);
+
+            var entities = _context.UserEntities;
+
+            if (id < 1 || !entities.Any() || id > entities.Max(u => u.Id)) {
+                throw new ArgumentException($"{id} is not a valid id.");
             }
 
+            UserEntity entity = await entities.FindAsync(id);
+            // assign all the values
+            {
+                entity.Status = userEntity.Status;
+                entity.FirstName = userEntity.FirstName;
+                if (!userEntity.Email.IsNullOrEmpty())
+                    entity.Email = userEntity.Email;
+                entity.LastName = userEntity.LastName;
+            }
+            // save changes.
+            _context.SaveChanges();
+            return true;
         }
 
         // id is the follower, userId is followee aka person that the follower is following
